* Added Test_Framework::event method   * 1) Checks to see if a fail_condition is met   * 2) Checks to see if a continue_condition is met   * 3) Waits on a pre_condition   * 4) Sets a post_condition   * 5) (Optional) Allows barriers on an event   * 6) Sets local variables that may be logged later * Added skeletons for log and dump methods   * Will be implemented soon * Changed entire KaRL engine to be long long   * will allow for valid timestamps for knowledge   * some printouts will have to be fixed (ACE_DEBUG %d changed to std::stringstream)
git-svn-id: http://madara.googlecode.com/svn/trunk@242 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ madara/Globals.h
@@ -48,8 +48,8 @@ namespace Madara
     {}
 
     int status;
-    long value;
-    unsigned long clock;
+    long long value;
+    unsigned long long clock;
     int scope;
 
     // for keeping track of current high priority originator
