+++ madara/expression_tree/Leaf_Node.h
@@ -22,7 +22,7 @@ namespace Madara
     {
     public:
       /// Ctor.
-      Leaf_Node (int item);
+      Leaf_Node (long long item);
 
       /// Ctor.
       Leaf_Node (const ::std::string &item);
@@ -34,23 +34,23 @@ namespace Madara
       virtual ~Leaf_Node (void);
 
       /// Return the item stored in the node.
-      virtual int item (void) const;
+      virtual long long item (void) const;
 
       /// Prune the tree of unnecessary nodes. 
       /// Returns evaluation of the node and sets can_change appropriately.
       /// if this node can be changed, that means it shouldn't be pruned.
-      virtual int prune (bool & can_change);
+      virtual long long prune (bool & can_change);
 
       /// Evaluates the node and its children. This does not prune any of
       /// the expression tree, and is much faster than the prune function
-      virtual int evaluate (void);
+      virtual long long evaluate (void);
 
       /// Define the @a accept() operation used for the Visitor pattern.
       virtual void accept (Visitor &visitor) const; 
 
     private:
       /// Integer value associated with the operand.
-      int item_;
+      long long item_;
     };
 
   }
